Split s3-related code on several files
