+++ src/main/scala/ohnosequences/awstools/s3/client.scala
@@ -20,155 +20,8 @@ import scala.concurrent.duration.Duration
 import scala.util.{Failure, Success, Try}
 
 import java.io.{IOException, InputStream, ByteArrayInputStream, File}
-import java.net.URI
 import java.net.URL
 
-
-sealed trait AnyS3Address {
-  // These are the inputs
-  val _bucket: String
-  val _key: String
-
-  def toURI: URI = new URI("s3", _bucket, s"/${_key}", null).normalize
-
-  // We use URI as a way to sanitize things (dropping extra /)
-  lazy val bucket: String = toURI.getHost
-  lazy val key: String = toURI.getPath
-
-  lazy val segments: Seq[String] = key.split("/").filter(_.nonEmpty).toSeq
-
-  @deprecated("Use toURI method instead, or just toString", since = "v0.17.0")
-  final def url = "s3://" + bucket + "/" + key
-
-  override def toString = toURI.toString
-
-  def toHttpsURL(region: Region): URL = new URL("https", s"s3-${region}.amazonaws.com", s"${bucket}/${key}")
-}
-
-case class S3Folder(b: String, k: String) extends AnyS3Address {
-  val _bucket = b
-  // NOTE: we explicitly add / in the end here (it represents the empty S3 object of the folder)
-  val _key = k + "/"
-
-  def /(suffix: String): S3Object = S3Object(bucket, key + suffix)
-}
-
-object S3Folder {
-
-  def apply(uri: URI): S3Folder = S3Folder(uri.getHost, uri.getPath)
-
-  implicit def toS3Object(f: S3Folder): S3Object =
-    S3Object(f.bucket, f.key.stripSuffix("/"))
-}
-
-
-case class S3Object(_bucket: String, _key: String) extends AnyS3Address {
-
-  def /(): S3Folder = S3Folder(bucket, key)
-
-  def /(suffix: String): S3Object = this./ / suffix
-}
-
-object S3Object {
-
-  def apply(uri: URI): S3Object = S3Object(uri.getHost, uri.getPath)
-}
-
-
-case class S3AddressFromString(val sc: StringContext) extends AnyVal {
-
-  // This allows to write things like s3"bucket" / "foo" / "bar" /
-  // or s3"org.com/${suffix}/${folder.getName}" / "file.foo"
-  def s3(args: Any*): S3Folder = {
-    val str = sc.s(args: _*)
-    S3Folder(new URI("s3://" + str))
-  }
-}
-
-
-
-case class TransferListener(transfer: Transfer) extends PListener {
-  def progressChanged(progressEvent: PEvent) {
-    import ProgressEventType._
-    progressEvent.getEventType match {
-      case TRANSFER_STARTED_EVENT  => println("Started")
-      case TRANSFER_CANCELED_EVENT  => println("Canceled!")
-      case TRANSFER_COMPLETED_EVENT  => println("Completed!")
-      case TRANSFER_FAILED_EVENT  => println("Failed!")
-      case TRANSFER_PART_COMPLETED_EVENT  => println("Completed part: "+ transfer.getProgress.getBytesTransferred)
-      case TRANSFER_PART_FAILED_EVENT  => println("Failed part transfer")
-      case TRANSFER_PART_STARTED_EVENT  => println("Started part transfer")
-      case TRANSFER_PREPARING_EVENT  => println("Preparing for the transfer")
-      // case HTTP_REQUEST_CONTENT_RESET_EVENT  => ()
-      case _ => ()
-    }
-  }
-}
-
-case class LoadingManager(transferManager: TransferManager, logger: Option[Logger] = None) {
-
-  val transferWaiter: (Transfer => Unit) = { transfer =>
-    while(!transfer.isDone) {
-      println(" - Progress: " + transfer.getProgress.getBytesTransferred + " bytes")
-      Thread.sleep(500)
-    }
-    println("Finished: " + transfer.getState)
-  }
-
-  // Asinchronous progress listener
-  val transferListener: (Transfer => Unit) = { transfer =>
-    transfer.addProgressListener(TransferListener(transfer))
-  }
-
-  def upload(
-    objectAddress: S3Object,
-    file: File,
-    transferWaiter: (Transfer => Unit) = transferWaiter
-  ) {
-    println("Uploading to: " + objectAddress.toString)
-    println("File: " + file.getAbsolutePath)
-    val upload = transferManager.upload(objectAddress.bucket, objectAddress.key, file)
-    transferWaiter(upload)
-  }
-
-  def uploadDirectory(
-    objectAddress: S3Object,
-    directory: File,
-    recursively: Boolean = true,
-    transferWaiter: (Transfer => Unit) = transferWaiter
-  ) {
-    println("Uploading to: " + objectAddress.toString)
-    println("Directory: " + directory.getCanonicalPath + (if (recursively) " (recursively)" else ""))
-    val upload = transferManager.uploadDirectory(objectAddress.bucket, objectAddress.key, directory, recursively)
-    transferWaiter(upload)
-  }
-
-  def download(
-    objectAddress: S3Object,
-    file: File,
-    transferWaiter: (Transfer => Unit) = transferWaiter
-  ) {
-    println("Dowloading from: " + objectAddress.toString)
-    println("File: " + file.getAbsolutePath)
-    val download = transferManager.download(objectAddress.bucket, objectAddress.key, file)
-    transferWaiter(download)
-  }
-
-  def downloadDirectory(
-    objectAddress: S3Object,
-    destinationDir: File,
-    transferWaiter: (Transfer => Unit) = transferWaiter
-  ) {
-    println("Dowloading from: " + objectAddress.toString)
-    println("To directory: " + destinationDir.getAbsolutePath)
-    val download = transferManager.downloadDirectory(objectAddress.bucket, objectAddress.key, destinationDir)
-    transferWaiter(download)
-  }
-
-  def shutdown() {
-    transferManager.shutdownNow(false)
-  }
-}
 
 class S3(val s3: AmazonS3) {
